+++ examples/document_verification.rb
@@ -35,7 +35,8 @@ image_details = [
     image_type_id: '<0 | 2>',
     image: '<full path to selfie image or base64image string>'
   },
-  { # Not required if you don't require proof of life (note photo of photo check will still be performed on the uploaded selfie)
+  { # Not required if you don't require proof of life (note photo of photo check
+    # will still be performed on the uploaded selfie)
     image_type_id: '<4 | 6>',
     image: '<full path to liveness image or base64 image string>'
   },
@@ -57,9 +58,15 @@ id_info = {
 
 # Set options for the job
 options = {
-  return_job_status: '<true | false>', # Set to true if you want to get the job result in sync (in addition to the result been sent to your callback). If set to false, result is sent to callback url only.
-  return_history: '<true | false>', # Set to true to receive all of the updates you would otherwise have received in your callback as opposed to only the final result. You must set return_job_status to true to use this flag.
-  return_image_links: '<true |false>', # Set to true to receive links to the selfie and the photo it was compared to. You must set return_job_status to true to use this flag.
+  # Set to true if you want to get the job result in sync (in addition to the result
+  # been sent to your callback). If set to false, result is sent to callback url only.
+  return_job_status: '<true | false>',
+  # Set to true to receive all of the updates you would otherwise have received in your
+  # callback as opposed to only the final result. You must set return_job_status to true to use this flag.
+  return_history: '<true | false>',
+  # Set to true to receive links to the selfie and the photo it was compared to.
+  # You must set return_job_status to true to use this flag.
+  return_image_links: '<true |false>',
   signature: true
 }