+++ examples/example-project/smart_bank.rb
@@ -10,8 +10,10 @@ class SmartBank
   attr_reader :partner_id, :default_callback, :api_key, :sid_server, :user_id, :job_id
 
   def initialize
-    @partner_id = ENV['SMILE_PARTNER_ID'] # login to the Smile Identity portal to view your partner id
-    @default_callback = ENV['SMILE_JOB_CALLBACK_URL'] # See https://docs.smileidentity.com/server-to-server/ruby/products/biometric-kyc#create-a-callback-endpoint
+    # login to the Smile Identity portal to view your partner id
+    @partner_id = ENV['SMILE_PARTNER_ID']
+    # See https://docs.smileidentity.com/server-to-server/ruby/products/biometric-kyc#create-a-callback-endpoint
+    @default_callback = ENV['SMILE_JOB_CALLBACK_URL']
     @api_key = ENV['SMILE_API_KEY'] # copy your API key from the Smile Identity portal
     @sid_server = ENV['SMILE_SERVER_ENVIRONMENT'] # Use '0' for the sandbox server, use '1' for production server
     @user_id = SecureRandom.uuid # your unique ID for the user
@@ -52,18 +54,14 @@ class SmartBank
   # Returns a verification result Hash
   def perform_biometric_kyc
     # Create image list
-    # image_type_id: Integer
-    # 0 - Selfie image jpg or png (if you have the full path of the selfie)
-    # 2 - Selfie image jpg or png base64 encoded (if you have the base64image string of the selfie)
-    # 4 - Liveness image jpg or png (if you have the full path of the liveness image)
-    # 6 - Liveness image jpg or png base64 encoded (if you have the base64image string of the liveness image)
     image_details = [
       {
-        image_type_id: 0,
+        image_type_id: SmileIdentityCore::ImageType::SELFIE_IMAGE_FILE,
         image: '/path/to/selfie_image.jpeg'
       },
-      { # Not required if you don't require proof of life (note photo of photo check will still be performed on the uploaded selfie)
-        image_type_id: 4,
+      { # Not required if you don't require proof of life (note photo of photo check will
+        # still be performed on the uploaded selfie)
+        image_type_id: SmileIdentityCore::ImageType::LIVENESS_IMAGE_FILE,
         image: '/path/to/liveness_image.jpeg'
       }
     ]
@@ -72,7 +70,7 @@ class SmartBank
     id_info = {
       first_name: 'John',
       last_name: 'Doe',
-      country: 'GH', # 2-letter country code>
+      country: 'GH', # 2-letter country code
       id_type: 'PASSPORT',
       id_number: 'G0000000',
       dob: '1992-12-07', # yyyy-mm-dd
@@ -90,30 +88,22 @@ class SmartBank
   # Returns a verification result Hash
   def perform_document_verification
     # Create image list
-    # image_type_id: Integer
-    # 0 - Selfie image jpg or png (if you have the full path of the selfie)
-    # 2 - Selfie image jpg or png base64 encoded (if you have the base64image string of the selfie)
-    # 4 - Liveness image jpg or png (if you have the full path of the liveness image)
-    # 6 - Liveness image jpg or png base64 encoded (if you have the base64image string of the liveness image)
-    # 1 - Front of ID document image jpg or png (if you have the full path of the selfie)
-    # 3 - Front of ID document image jpg or png base64 encoded (if you have the base64image string of the selfie)
-    # 5 - Back of ID document image jpg or png (if you have the full path of the selfie)
-    # 7 - Back of ID document image jpg or png base64 encoded (if you have the base64image string of the selfie)
     image_details = [
       {
-        image_type_id: 0,
+        image_type_id: SmileIdentityCore::ImageType::SELFIE_IMAGE_FILE,
         image: '/path/to/selfie_image.jpeg'
       },
-      { # Not required if you don't require proof of life (note photo of photo check will still be performed on the uploaded selfie)
-        image_type_id: 4,
+      { # Not required if you don't require proof of life (note photo of photo
+        # check will still be performed on the uploaded selfie)
+        image_type_id: SmileIdentityCore::ImageType::LIVENESS_IMAGE_FILE,
         image: '/path/to/liveness_image.jpeg'
       },
       {
-        image_type_id: 1,
+        image_type_id: SmileIdentityCore::ImageType::ID_CARD_IMAGE_FILE,
         image: '/path/to/front_document_image.jpeg'
       },
       { # Optional, only use if you're uploading the back of the id document image
-        image_type_id: 5,
+        image_type_id: SmileIdentityCore::ImageType::ID_CARD_BACK_IMAGE_FILE,
         image: '/path/to/back_document_image.jpeg'
       }
     ]
@@ -138,22 +128,18 @@ class SmartBank
     partner_params = {
       user_id: '512c9c37-a689-4959-a620-bed75fb41344', # previously registered user's user_id
       job_id: SecureRandom.uuid, # new unique job ID
-      job_type: 2
+      job_type: SmileIdentityCore::JobType::SMART_SELFIE_REGISTRATION
     }
 
     # Create image list
-    # image_type_id: Integer
-    # 0 - Selfie image jpg or png (if you have the full path of the selfie)
-    # 2 - Selfie image jpg or png base64 encoded (if you have the base64image string of the selfie)
-    # 4 - Liveness image jpg or png (if you have the full path of the liveness image)
-    # 6 - Liveness image jpg or png base64 encoded (if you have the base64image string of the liveness image)
     image_details = [
       {
-        image_type_id: 0,
+        image_type_id: SmileIdentityCore::ImageType::SELFIE_IMAGE_FILE,
         image: '/path/to/selfie_image.jpeg'
       },
-      { # Not required if you don't require proof of life (note photo of photo check will still be performed on the uploaded selfie)
-        image_type_id: 4,
+      { # Not required if you don't require proof of life
+        # (note photo of photo check will still be performed on the uploaded selfie)
+        image_type_id: SmileIdentityCore::ImageType::LIVENESS_IMAGE_FILE,
         image: '/path/to/liveness_image.jpeg'
       }
     ]
@@ -180,9 +166,15 @@ class SmartBank
 
   def job_options
     {
-      return_job_status: true, # Set to true if you want to get the job result in sync (in addition to the result been sent to your callback). If set to false, result is sent to callback url only.
-      return_history: false, # Set to true to receive all of the updates you would otherwise have received in your callback as opposed to only the final result. You must set return_job_status to true to use this flag.
-      return_image_links: true, # Set to true to receive links to the selfie and the photo it was compared to. You must set return_job_status to true to use this flag.
+      # Set to true if you want to get the job result in sync (in addition to the result been sent to your callback).
+      # If set to false, result is sent to callback url only.
+      return_job_status: true,
+      # Set to true to receive all of the updates you would otherwise have received in your callback as opposed to
+      # only the final result. You must set return_job_status to true to use this flag.
+      return_history: false,
+      # Set to true to receive links to the selfie and the photo it was compared to.
+      # You must set return_job_status to true to use this flag.
+      return_image_links: true,
       signature: true
     }
   end
@@ -195,7 +187,8 @@ enhanced_kyc_response['success'] # => true
 enhanced_kyc_response['result']['PartnerParams']['job_id'] # job_id
 enhanced_kyc_response['result']['PartnerParams']['user_id'] # user_id
 enhanced_kyc_response['result']['PartnerParams']['job_type'] # => 5
-# See https://docs.smileidentity.com/products/for-individuals-kyc/identity-lookup#return-values for the full JSON response interpretation
+# See https://docs.smileidentity.com/products/for-individuals-kyc/identity-lookup#return-values
+# for the full JSON response interpretation
 
 # Biometric KYC
 bio_kyc_response = smart_bank.perform_biometric_kyc